Fix for fire/water race conditions

+++ aegir/tools/host/host-fire.sh
@@ -7,45 +7,45 @@ guest_guard() {
 if [ ! -e "/var/run/fire.pid" ] && [ ! -e "/var/run/water.pid" ]; then
   touch /var/run/fire.pid
   for i in `dir -d /vservers/*`; do
-    if [ -e "$i/var/xdrago/monitor/ssh.log" ]; then
+    if [ -e "$i/var/xdrago/monitor/ssh.log" ] && [ -e "/usr/var/run${i}" ]; then
       for _IP in `cat $i/var/xdrago/monitor/ssh.log | cut -d '#' -f1 | sort`; do
         _FW_TEST=$(csf -g ${_IP} 2>&1)
         if [[ "${_FW_TEST}" =~ "DENY" ]] || [[ "${_FW_TEST}" =~ "ALLOW" ]]; then
           echo "${_IP} already denied or allowed on port 22"
         else
           echo "Deny ${_IP} on ports 21,22,443,80 in the next 1h"
-          csf -td ${_IP} 3600 -p 21
-          csf -td ${_IP} 3600 -p 22
-          csf -td ${_IP} 3600 -p 443
-          csf -td ${_IP} 3600 -p 80
+          csf -td ${_IP} 900 -p 21
+          csf -td ${_IP} 900 -p 22
+          csf -td ${_IP} 900 -p 443
+          csf -td ${_IP} 900 -p 80
         fi
       done
     fi
-    if [ -e "$i/var/xdrago/monitor/web.log" ]; then
+    if [ -e "$i/var/xdrago/monitor/web.log" ] && [ -e "/usr/var/run${i}" ]; then
       for _IP in `cat $i/var/xdrago/monitor/web.log | cut -d '#' -f1 | sort`; do
         _FW_TEST=$(csf -g ${_IP} 2>&1)
         if [[ "${_FW_TEST}" =~ "DENY" ]] || [[ "${_FW_TEST}" =~ "ALLOW" ]]; then
           echo "${_IP} already denied or allowed on port 80"
         else
           echo "Deny ${_IP} on ports 21,22,443,80 in the next 1h"
-          csf -td ${_IP} 3600 -p 21
-          csf -td ${_IP} 3600 -p 22
-          csf -td ${_IP} 3600 -p 443
-          csf -td ${_IP} 3600 -p 80
+          csf -td ${_IP} 900 -p 21
+          csf -td ${_IP} 900 -p 22
+          csf -td ${_IP} 900 -p 443
+          csf -td ${_IP} 900 -p 80
         fi
       done
     fi
-    if [ -e "$i/var/xdrago/monitor/ftp.log" ]; then
+    if [ -e "$i/var/xdrago/monitor/ftp.log" ] && [ -e "/usr/var/run${i}" ]; then
       for _IP in `cat $i/var/xdrago/monitor/ftp.log | cut -d '#' -f1 | sort`; do
         _FW_TEST=$(csf -g ${_IP} 2>&1)
         if [[ "${_FW_TEST}" =~ "DENY" ]] || [[ "${_FW_TEST}" =~ "ALLOW" ]]; then
           echo "${_IP} already denied or allowed on port 21"
         else
           echo "Deny ${_IP} on ports 21,22,443,80 in the next 1h"
-          csf -td ${_IP} 3600 -p 21
-          csf -td ${_IP} 3600 -p 22
-          csf -td ${_IP} 3600 -p 443
-          csf -td ${_IP} 3600 -p 80
+          csf -td ${_IP} 900 -p 21
+          csf -td ${_IP} 900 -p 22
+          csf -td ${_IP} 900 -p 443
+          csf -td ${_IP} 900 -p 80
         fi
       done
     fi
@@ -57,28 +57,15 @@ fi
 
 if [ -e "/vservers" ] \
   && [ -e "/etc/csf/csf.deny" ] \
+  && [ ! -e "/var/run/water.pid" ] \
   && [ -e "/usr/sbin/csf" ]; then
-  guest_guard
-  sleep 5
-  guest_guard
-  sleep 5
-  guest_guard
-  sleep 5
-  guest_guard
-  sleep 5
-  guest_guard
-  sleep 5
-  guest_guard
-  sleep 5
-  guest_guard
-  sleep 5
-  guest_guard
-  sleep 5
-  guest_guard
-  sleep 5
-  guest_guard
-  sleep 5
-  guest_guard
+  [ ! -e "/var/run/water.pid" ] && guest_guard
+  sleep 15
+  [ ! -e "/var/run/water.pid" ] && guest_guard
+  sleep 15
+  [ ! -e "/var/run/water.pid" ] && guest_guard
+  sleep 15
+  [ ! -e "/var/run/water.pid" ] && guest_guard
   rm -f /var/run/fire.pid
 fi
 exit 0
